ReadElf: Read ELF header for 32-bit and 64-bit ELF files
Provide initial documentation for design and create a simple skeleton of code based on that design.

While it looks like a very large patch, most of it is in test cases, that read files containing the first 64-bytes of an ELF file. I've tried to capture as many "useful" ELF files as I could, They're proprietary binaries, so they full contents can't be distributed, only the header information which contains no source code or other copyright information.

Issue: RUST-2