ReadElf: Implement reading from a file
Allow reading from a file on disk. This allows interpreting ELF files that are larger than the native size on 32-bit machines, by not having to load the file into memory first.

We use a `File` directly for reading, without wrapping it around a `BufReader`, as there are a lot of seeks when reading an ELF file. According to Rust documentation [1], a seek will discard the internal buffer, so there wouldn't be any real benefits. The existing code would have to be profiled and tested when reading from a file to make real benefits for performance increases.

Issue: RUST-3

--
[1] https://doc.rust-lang.org/std/io/struct.BufReader.html#method.seek