ReadElf: Iterate over program headers
Use `ReadElf::program_headers(&self)` to get an iterator to know all the program headers and their segment type. Iteration is done lazily. The struct `ProgramHeaders` is returned which is an iterator. The method `len()` can be used to get the `e_phnum` entries.

The binary from the crate (main.rs) has been moved now to an example, so build instructions are updated.

Issue: RUST-4