ReadElf: Allow to get the index of a ProgramHeader
Provide a method `index(index: usize)` on the `ProgramHeaders` struct. The `usize` is chosen to be consistent with indexing types, even though it cannot be anything other than a `u16`.

The trait `std::ops::Index` (or `std::ops::Deref`) could not be implemented because we must return a reference to self, where in this implementation we need to make a new instance of the struct.

Issue: RUST-6