[WIP] ReadElf: Add design for section headers
Provide iterators to get all the sections in the ELF file, via `ReadElf::section_headers()`.

Extend the `BinParser` trait with a `get_map()` that returns a block of memory within the ELF file. For files, it reads it from storage into a `Vec<u8>`, otherwise, it's just an in-place reference. Test cases extended to check the functionality.

Remove inline for `BinParser` as it doesn't make sense (inlining via the vtable from a trait isn't possible). We only want to inline `Slice` as we use this directly from `VecBuffer`.

TODO:
- Unit tests
  - Write to a string section
  - Create and test various paths

Issue: RUST-7